tests: refactor TestEventPayload in a dedicated file
